requirements.txt
apache-airflow==3.1.3
certifi==2025.10.5
Flask==3.1.2
joblib==1.5.2
mlflow==3.3.2
mlflow_skinny==3.3.2
mlflow_tracing==3.3.2
nltk==3.9.1
numpy==2.3.5
pandas==2.3.3
PyYAML==6.0.3
scikit_learn==1.7.2
scipy==1.16.3
textblob==0.19.0
